requirements.txt
apturl==0.5.2
asttokens==2.4.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
diffusers==0.25.0
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
future==0.18.2
httplib2==0.20.2
huggingface-hub==0.20.1
idna==3.3
importlib-metadata==4.6.4
invisible-watermark==0.2.0
ipython==8.20.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
packaging==23.2
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycairo==1.20.1
pycups==2.0.1
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.30
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python3-xlib==0.15
PyTMX==3.32
pytweening==1.0.7
pytz==2022.1
PyWavelets==1.5.0
pyxdg==0.27
PyYAML==5.4.1
regex==2023.12.25
reportlab==3.6.8
requests==2.25.1
safetensors==0.4.1
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.12
systemd-python==234
terminator==2.1.1
tokenizers==0.12.1
torch==2.1.2
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
traitlets==5.14.1
transformers==4.19.2
triton==2.1.0
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
zipp==1.0.0
